feat(agent-tars): refine `browser_get_markdown` implementation
